api/lineq-solver/requirements.txt
azure.functions
azure-functions-worker
numpy